+++ crates/backend-uzu/src/encodable_block/dflash.rs
@@ -170,8 +170,8 @@ impl<B: Backend> DFlash<B> {
         })
     }
 
-    pub fn block_size(&self) -> usize {
-        self.block_size as usize
+    pub fn block_size(&self) -> u32 {
+        self.block_size
     }
 
     pub fn empty_state(
